Update TAG Generation Logic
The Tag generation logic currently results in the image having a tag like the following:

release-v1.30.0-v.1.30.0

This is as we use both the release branch name and the tag name. What we have previously done
to get around this is manually edit the tags once pushed up to Quay. This commit introduces a change
such that the generated tag matches the release branch exactly.

Signed-off-by: Anish Asthana <anishasthana1@gmail.com>